Backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==21.2.0

mysql-connector-python==9.0.0

pydantic==2.10.4
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3

bcrypt==4.1.3
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-jose==3.3.0

# ===== BLOCO SUPABASE COMPATÍVEL PYTHON 3.13 =====
supabase==2.3.4
gotrue==2.4.2
postgrest==0.15.1
storage3==0.7.5
realtime==1.0.6
httpx==0.25.2

mercadopago==2.2.3
google-auth==2.34.0

reportlab==4.2.2
fpdf2==2.7.9
python-barcode==0.15.1
pillow==10.4.0
qrcode[pil]==7.4.2

cryptography==41.0.7
stripe==10.8.0